+++ management/framework.php
@@ -45,28 +45,12 @@ $framework = $DB->get_record('tool_mutrain_framework', ['id' => $id], '*', MUST_
 $context = context::instance_by_id($framework->contextid);
 require_capability('tool/mutrain:viewframeworks', $context);
 
-$pageurl = new \moodle_url('/admin/tool/mutrain/management/framework.php', ['id' => $framework->id]);
+$pageurl = new \core\url('/admin/tool/mutrain/management/framework.php', ['id' => $framework->id]);
 
 management::setup_framework_page($pageurl, $context, $framework);
 
 /** @var \tool_mutrain\output\management\renderer $managementoutput */
 $managementoutput = $PAGE->get_renderer('tool_mutrain', 'management');
-
-$actions = new header_actions(get_string('management_framework_actions', 'tool_mutrain'));
-if (has_capability('tool/mutrain:manageframeworks', $context)) {
-    if (\tool_mutrain\local\framework::is_deletable($framework->id)) {
-        $url = new moodle_url('/admin/tool/mutrain/management/framework_delete.php', ['id' => $framework->id]);
-        $link = new tool_mulib\output\ajax_form\link($url, get_string('framework_delete', 'tool_mutrain'));
-        $link->set_submitted_action($link::SUBMITTED_ACTION_REDIRECT);
-        $actions->get_dropdown()->add_ajax_form($link);
-    }
-    $url = new \moodle_url('/admin/tool/mutrain/management/framework_update.php', ['id' => $framework->id]);
-    $button = new \tool_mulib\output\ajax_form\button($url, get_string('framework_update', 'tool_mutrain'));
-    $actions->add_button($button);
-}
-if ($actions->has_items()) {
-    $PAGE->add_header_action($OUTPUT->render($actions));
-}
 
 echo $OUTPUT->header();
 
@@ -78,7 +62,7 @@ $table = new \tool_mutrain\table\fields($pageurl, $framework);
 $table->out($table->pagesize, false);
 
 if (!$framework->archived && has_capability('tool/mutrain:manageframeworks', $context)) {
-    $url = new \moodle_url('/admin/tool/mutrain/management/field_add.php', ['frameworkid' => $framework->id]);
+    $url = new \core\url('/admin/tool/mutrain/management/field_add.php', ['frameworkid' => $framework->id]);
     $button = get_string('field_add', 'tool_mutrain');
     $button = new \tool_mulib\output\ajax_form\button($url, $button);
     $addbutton = $OUTPUT->render($button);