+++ management/framework_delete.php
@@ -17,7 +17,7 @@
 // phpcs:disable moodle.Files.BoilerplateComment.CommentEndedTooSoon
 
 /**
- * Delete training framework.
+ * Delete credit framework.
  *
  * @package    tool_mutrain
  * @copyright  2024 Open LMS (https://www.openlms.net/)
@@ -30,8 +30,6 @@ use tool_mutrain\local\framework;
 
 /** @var moodle_database $DB */
 /** @var moodle_page $PAGE */
-/** @var core_renderer $OUTPUT */
-/** @var stdClass $CFG */
 
 define('AJAX_SCRIPT', true);
 
@@ -45,11 +43,11 @@ $framework = $DB->get_record('tool_mutrain_framework', ['id' => $id], '*', MUST_
 $context = context::instance_by_id($framework->contextid);
 require_capability('tool/mutrain:manageframeworks', $context);
 
-$currenturl = new moodle_url('/admin/tool/mutrain/management/framework_delete.php', ['id' => $framework->id]);
+$currenturl = new core\url('/admin/tool/mutrain/management/framework_delete.php', ['id' => $framework->id]);
 $PAGE->set_context($context);
 $PAGE->set_url($currenturl);
 
-$returnurl = new moodle_url('/admin/tool/mutrain/management/index.php', ['contextid' => $context->id]);
+$returnurl = new core\url('/admin/tool/mutrain/management/index.php', ['contextid' => $context->id]);
 
 if (!framework::is_deletable($framework->id)) {
     // We should not get here.